moved socksify to vault package
