+++ vault/socksify.go
@@ -1,4 +1,4 @@
-package auth
+package vault
 
 import (
 	"io/ioutil"
